+++ packages/g-plugin-svg-picker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-plugin-svg-picker
 
+## 1.9.14
+
+### Patch Changes
+
+-   Updated dependencies [c0b9fc49]
+    -   @antv/g-plugin-svg-renderer@1.10.14
+
 ## 1.9.13
 
 ### Patch Changes